Move RequestInfo to CallCredentials
