+++ core/src/test/java/io/grpc/internal/CallCredentials2ApplyingTest.java
@@ -30,8 +30,8 @@ import static org.mockito.Mockito.when;
 
 import io.grpc.Attributes;
 import io.grpc.CallCredentials.MetadataApplier;
+import io.grpc.CallCredentials.RequestInfo;
 import io.grpc.CallCredentials2;
-import io.grpc.CallCredentials2.RequestInfo;
 import io.grpc.CallOptions;
 import io.grpc.IntegerMarshaller;
 import io.grpc.Metadata;
